Add component midonet-brain and basic VTEP support
This commit includes a new submodule midonet-brain that will hold all
code related to the controller cluster. It includes a submodule with the
odl-ovsdb fork at a revision that is usable programatically, plus a very
simple implementation of a VtepDataClient that is able to connect to an
ovsdb instance using the odl libraries.

The operations implemented are:
- Add a logical switch
- Add a vlan-binding to a physical port
- Add mcast and ucast remote macs
- List physical and logical switches, ports, ucast and mcast macs

Some of these calls make certain assumptions about the VTEP
configuration, which are acceptable for the controlled demo we want to
target in this milestone.

Ref: MN-1542, MN-1548

Change-Id: I89237e2d40f101dfb1db6d721e8d5a47e8af5fec
Signed-off-by: Galo Navarro <galo@midokura.com>