Removing common translateFlowActions path
In translateAction(), the FlowTranslator carefully seperates traffic to
port sets from traffic to single ports, but all the towards*() functions
converge into the same common translateFlowActions() function for
emitting the final list of FlowActions, which induces some extra boxing
and extra logic to reseperate again in translateFlowActions().

This patch eliminates translationFlowActions() and move back the
specific logic needed in towardsLocalDpPorts(), towardsRemoteHosts()
and toPortSet(), renamed for consistency in towardsPortSet().

Change-Id: I40cb2ff73975345170bdb3c9a954f0fd39f40f37
Signed-off-by: hugobenichi <hugo@midokura.com>