Add the validation of mappingStatus to the VIP and PoolMember
This patch adds the check that the related pool's `mappingStatus` is not
`PENDING_CREATE`, `PENDING_DELETE` and `PENDING_UPDATE` to the methods
in `LocalDataClientImpl` for the VIP and PoolMember.

With the previous implementation even if the ZooKeeper manipulations
failed because of `mappingStatus` is invalid, the requested
create/update requests for the VIP or PoolMember are commited to
ZooKeeper but the following manipulations to create/update the
Pool-HealthMonitor mapping are failed and the API server returns 503
Service Unavailable. This means even if the server returns the error,
the requested resources can be created/updated.

This behaviour violates the consistency of the data in ZooKeeper and
this patch fixes it and make the API detect if `mappingStatus` is in the
valid value before it starts side-effective manipulations. The server
would return 503 Service Unavailable and the requested resources are not
going to be created/updated.

Change-Id: Ifbed6b4e2e10781da7b1f33cab26c9a17593633b
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>