Remove missleading log about lost userspace flows
This log statement does not make any sense anymore after the changes
introduced for the ICMP dp flow optimization (MN-1071) and the
associated code path was preventing missing flows to be forgotten
properly. By definition a flow added by midolman can't be a userspace
only flow, or it is in the first place a bug of the code taking the
decision to install or not a dp flow.

Change-Id: I58686d7725e4b949d3bf8ccd03dfa0d0536800e1
Signed-off-by: hugobenichi <hugo@midokura.com>