Ensures adequate number of tokens
This patches reverts 82235a116557d94cdec2c69699bc4d0650a2da02 and
instead ensures that a given bucket has enough tokens to
evenly distribute to its children. This is done by adjusting the
capacity of the root bucket when new leafs are added.

Also, it fixes a typo in doDistribution() that caused a
miscalculation and a token leak.

Signed-off-by: Duarte Nunes <duarte@midokura.com>
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>
Change-Id: Icf3bfb2176fc7889d7935cc76d2256dc11ef4b32