Ensure callbacks of NetlinkRequests are run once
In AbstractNetlinkConnection, the read and write threads could be
racing to activate the callback of a NetlinkRequest. NetlinkRequest
was checking that a single callback would not be activated twice, but
in commit 89a0c6e97942822ea8bf4000fd02786498c70fbc this was relaxed
and because of the above race, a callback could get run twice,
triggering wrong behaviour in the request clients.

This patch removes the race and ensure a the read and write threads
would not try to both activate the same callback, and it also makes
the NetlinkRequest check again that its callback is not run twice.

Ref MN-1617
Signed-off-by: hugobenichi <hugo@midokura.com>
Change-Id: Ie19a9161d21484eed35789b89f55b5b0021a8d91
Signed-off-by: hugo <hugo@midokura.jp>