Preseed macs upon vtep binding
Takes all known MACs in the port being bound, figures out the host ip
based on the binding and host's tunnel zone membership, and writes the
relevant entries to the UCast_Macs_Remote table in the VTEP.

No IPs will be added to the MCast_Macs_Remote with the unknown-dst
wildcard, because at the moment we have no way to chose an L2 flooding
proxy. This will be implemented at a later stage.

Also adds minor comment and log msg cleanup.

Change-Id: Iac422c430f802b6f4b96f5a2fab3f380df2fc44b
Signed-off-by: Galo Navarro <galo@midokura.com>