Updates a hard-coded script directory.
MMM/MDTS code have been moved to under the midonet repository. MDTS
tests rely on a utility function to look up the directory where the
test library exists but the directory prefix is hard-coded in the
script and needs to be updated to reflect the move.

Eventually the test directory prefix should be moved to a config file.

Fix QA-145

Change-Id: I4cc41fec9e7d6fc582fe6e9ba2568832f1b8c6f7