Updates a hard-coded script directory.
MMM/MDTS code have been moved to under the midonet repository. MDTS
tests rely on a utility function to look up the directory where the
test library exists but the directory prefix is hard-coded in the
script and needs to be updated to reflect the move.

Eventually the test directory prefix should be moved to a config file.

Fix QA-145

Change-Id: Ib4c8776b0edf0fa4dcfb4ffc62a7f8456f9d27c0