Add tunnel zone id to VTEP resource
This tunnel zone will be used to decide which of the host's IP is to be
used as VxLAN tunnel endpoint. Tunnel zone deletions will fail if there
is a VTEP configured with the same tunnel zone.

TestTunnelZone is *not* updated because we still don't have the vtep
client mock that allows creating vteps in tests without actually
connecting to the switch. A later patch will add that mock, and the
updated tunnel zone test.

Change-Id: I9d53a3190874ecb9d1d63d9fdcc0744948b3ab1d
Signed-off-by: Galo Navarro <galo@midokura.com>