Add method to get vxlan end point from vxlan port
To the LocalDataClientImpl. This will be used in further patches to
retrieve the ip used by a host as its own vxlan tunnel end point. The
method will be used to instruct the VTEP regarding which midolman host
it should send traffic addressed to a given MAC.

Untested because it requires an ovsdb mock that will be added in a later
patch. We will add the test for this new method in there.

Change-Id: I4c7195237871daf4341097aedcfdde7431a39f99
Signed-off-by: Galo Navarro <galo@midokura.com>