Make brain/odl-ovsdb a proper maven dep
This was out of the maven build so it required compiling and adding the
jar manually. Now this won't happen, although the project will require a
git submodule --init --recursive in order to build properly.

Also updates the odl-ovsdb submodule bringing latest dev up to delete
apis, since we must change the poms inside.

Change-Id: I636726e8b56552276db56bf3f5dff347ea0db5f3
Signed-off-by: Galo Navarro <galo@midokura.com>