HTB face-lifting
This patch refactors the distribution algorithm to be performed one
token at a time. This simplifies the logic and makes it more fair
with regards to buckets that concurrently become non-full and would
be assigned more tokens than if they had been considered non-full at
the beginning of the iteration. Distribution is now also confined to
one thread.

To amortize the costs of having more distribution iterations, we keep
a distribution-private counter at each bucket where we accumulate
that bucket's reserved tokens. At the end of the distribution, we
atomically add those reserved tokens to the running, thread-safe
counters.

We also introduce the idea of a token multiplier. Each token we
consume is now worth 8 tokens. This means that we don't have to
atomically consume tokens and contend with concurrent distributions,
which also get triggered less often. A problem arises if we read less
packets then the tokens we obtained. For that case we add the
remaining tokens to a special position in the StatisticalCounter. This
position is the same for all channels, so we need an atomic, possibly
contented operation in the one_to_one configuration. This is OK, because
the contending threads are associated with channels under low load.

This patch also adds a unit test framework to enable succinct
specification of HTB tests.

Signed-off-by: Duarte Nunes <duarte@midokura.com>
Change-Id: I9162fd06c9e1c79b1bbad5372da45989eceaa07b