VTPM: better handling of ZK errors
ZK errors for additions/deletions of port set membership were
generic. Meaning that non disconnection related errors were unhandled (this was
left for the ZkConnnectionWatcher to decide).

However, operations that succeed on the server side but fail on the client due
to an untimely disconnection are not idempotent. In those cases the VTPM was
forever left in an inconsistent state wrt that particular port set, which would
never get status updates again.

This patch implements handling of such errors:

  * For new memberships: NodeExists is treated as success.
  * For removed memberships: NoNode is treated as success.

Fixes: MN-1633.

Change-Id: Ie961376c795fb1d51f4ecb3cf95fdaf3029f5bb8
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>