Removing FlowAction list from packets
Attaching the list of flow actions to org.midonet.odp.Packet for
execution from user space was not matching the "packet" structure of
the datapath, and was inducing unnecessary parsing in deserialisation.
It also reduces the amount of mutability on Packet in the midolman
layer.

This patch is preparing for two additional improvements:
  - simplification of the "isUserspace" handling for special icmp
    flow keys and userspace flow translation.
  - avoiding scala list to java list copying in ovs flow operations
    and in packet execute operations.

Change-Id: I0145eabecf8c10bc7157c38377dd5cf1f0fd7daa
Signed-off-by: hugobenichi <hugo@midokura.com>