Changed flow controller invalidation history
Allow zero (0) as a valid last seen invalidation
event identifier. Previously, id=0 was a "special
value" indicating that any tag set should be reported
as valid. With this change, use a negative value
for the last seen invalidation event id as "special
value". When using 0 as the last seen event id,
the invalidation history reports if the specified
tag set appears in any invalidation event, starting
with id from 1. This allows the correct reporting
of the first invalidated tag set with id=1.

Change-Id: I63b71dde0202efc3e870fbd84898ba352af6c163