Fix local race in ARP table
If an ARP entry is resolved locally because an ARP reply is received by the
vrouter, midolman should not notify simulations that are waiting for that entry
until the entry is written to zookeeper and zookeeper has notified the local
replicated map. Otherwise, the restarts of those simulations may find a NULL
value and overwrite the inflight real-value.

Change-Id: I635a3a1e8bde07c4f007c98ea434898fd3d830aa
Fixes: MN-1746
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>