Netlink/ovs layers endianness cleanup
The NetlinkMessage and Builder helper classes for serialisation and
deserialisation of Netlink attributes were made generic to support
either LE or BE native order to talk to the ovs datapath over netlink.

Because these intermediary objects put GC pressure on both
serialisation and deserialisation, it is desirable to remove them,
while preserving the endian agnostic property of the code.

Instead, this patch makes the NetlinkMessage and Builder use the
implicit system order, removes all the "helper" functions that were
taking ByteOrder arguments, and let the users sort out the endianness.

Therefore in org.midonet.odp, it is assumed that ByteBuffers are always
in system order, and FlowKeys that have fields written in BE network
order in the datapath are read from or written to ByteBuffers with the
help of a new helper utility class called BytesUtil.

BytesUtil is an interface that exposes reverseBytes functions for
short, int and long values, as well as ByteBuffer allocation functions.
Two "singleton" implementations exists, one for LE systems and one for
BE systems. At startup, a check is performed and one of the
implementation is selected according to the value of
ByteOrder.nativeOrder().

This makes the code for NetlinkMessage less stateful (does not track
the byte order internally anymore), more JIT friendly, and prepares
for further serialisation/deserialisation simplification.

Please note that the #slice() method does not preserve endianness of
the original buffer. Since the default order on java is BE, every slices
need to have its order set to LE explicitly when the system order is LE.

Change-Id: I3fe62fd2a6f3cdba204fbe59ad4a328366bda53b
Signed-off-by: hugobenichi <hugo@midokura.com>