ipfrag test: allow time for temp flows to expire
Drop flows for ip fragmentation are temporary because we want flows to
come to userspace once in a while, it give MM the chance to send ICMP
fragmentation needed messages.

When a test case leaves such a flow installed, the flow can pill over
to the next test. This method prevents that.

The wait time that guarantees that flows will be ejected is 15 seconds.
5 seconds is their hard expiration time, and 10 seconds is the flow
expiration check interval in the flow controller.

This test introduces a 15 second wait time before each ipfrag test case.

Change-Id: Ib83dfde3a16e292886cf5859b78d9d290f11cb26
Fixes: MN-1758
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>