Add basic pieces of a vxgw service
These are the basic classes that define a VxlanGatewayService. It is
composed of VxLanPeers that abstract a particular implementation (e.g.:
a physical switch, or a Midonet cloud), offering an rxJava Observable of
updates on the local macs, and allowing to apply updates on remote macs
reported by other VxLanPeers.

This commit also adds rxJava as a new dependency in the brain project,
which is used for observable streams exposed by VxLanPeers.

Ref. MN-1636
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I6c81c5bbf809b7bccf82e856df2cc3afd0583d5e