Enrich VtepDataClient for vxlan gw service
Includes new methods required to expose updates as observables, and
apply remote changes.

Moved logic from the VtepResource to construct the logical switch name
from a bridge id.  This is now in the VtepDataClient because the names
are standard and should be respected by any implementation of the data
client.

Ref. MN-1636
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ic59ccb2053b46e61a8f6689a252aa08b39f8d510