Implement VxLanPeer for Midonet: MAC stream
This patch provides an implementation of a VxLanPeer which listens on
mac tables for the set of bridges bound to vteps, and exposes these as
an Observable. This patch is not complete and will not apply the macs
reported from the other VxLanPeer, it simply provides a skeleton
exposing the MAC-port updates.

Also refactors older classes, adds tests, and includes a new method in
the DataClient to allow fetching vxlan bond ports.

Change-Id: Icb69ced6ed0ab6277390183e1047871cd8e6f725
Ref: MN-1636
Signed-off-by: Galo Navarro <galo@midokura.com>