VxLanMapper actor manages vni 2 uuid map
A new VxLanMapper actor is introduced as a child of the VTPM, which
spawn that new actor at startup.

The VxLanMapper acts as a simple bridge between the PacketWorkflow and
Zk and maintains a vni to uuid map that allows the PacketWorkflow to
handle ingress tunneled traffic coming from vtep peers corretly.

The VxLanMapper when it starts sends a query to Zk to get the list of
active vxlan ports ids. From this list it will then query the VTA and
construct the vni to uuid map. The VxLanMapper gets notified about
change to the list of vxlan ports ids via a watcher. The map itself
is holded for the time beeing in the actor companion object, which
allows synchronous access from the PacketWorkflow.

Fix MN-1747
Signed-off-by: hugobenichi <hugo@midokura.com>
Change-Id: I748c4bc32a56cbe1a5b6daa2ac9f3af83e33e41a