Removing Result wrapping in Zk Directory callbacks
While writing the patch for MN1747 and having to look at the Zk
directory callback "framework", I found out that the Result wrapping
was unnecessary since the Stat object was never used and always set to
null. This patch removes it and removes the associated complexity in
the directory callback api.

Change-Id: I534f5c07cca5b23da17e7d6a92fadf983cf4c15d
Signed-off-by: hugobenichi <hugo@midokura.com>