VtepBroker: handle apply and feed observable
This commit adds meat into the VtepBroker so that it's able to feed
events into its observable with notifications received from the VTEP,
and uses new code in the ovsdb client to delete entries from the ucast
mac remotes table.

This adds tests for vtep update translation, and also fixes various bugs
found in this code path.

Change-Id: I51514f005215c939bd9dc8e998a8df735515cb89
Signed-off-by: Galo Navarro <galo@midokura.com>