[MN-1811] Adds back a MacLearningTable notify call.
Fix a regression introduced in:
https://reviews.midonet.org:8443/#/c/1980

The patch changed a location for calling MacLearningTable.notify()
from within a BridgeBuilder to ClusterBridgeManager in order to let
ClusterBridgeManager handles all the callback management and to let
BridgeBuilder focus on constructing bridge topology data. However, the
patch missed a place where MacLearningTable is created for an unseen
VLAN ID of a bridge.

This patch introduced a helper function to set up a MacLearningTable
with a bridge, and made the code in question call the helper function.

Fix MN-1811

Change-Id: I9181db074134d646e0e5418bea1385d36cd57b81