OVS request Netlink serialisation
This patch removes the Builder and NetlinkMessage intermediary objects
in all flow requests and in packet execute requests, which are the vast
majority of datapah requests in midolman.

For this purpose, serialisation based on org.midonet.netlink.Translator
is added to the FlowKey and FlowAction classes hierarchy, with a couple
of additional helper methods in NetlinkMessage.

Change-Id: Ibac0ecf1e0df9e5c5ba1a35f3498e271b3df17c1
Signed-off-by: hugobenichi <hugo@midokura.com>