VXGW flooding proxy: add flooding weight to hosts
This is part of the flooding proxy implementation. This patch
adds a new field into the Host entity that allows specifying
a weight to be used by the controller to decide which midolman
host should be getting the flooded packets from a vtep.
This patch does not include the monitoring of changes of the value,
nor the algorithm to take the actual decision (these will be added
in a different patch).

Ref: MN-1825

Change-Id: I028d7a432c0f80d7845a9c75b31d008733dd3cc2
Signed-off-by: Ernest Artiaga <ernest@midokura.com>