Remove AttrKey wrapper around NL attribute ids
This class was now only used in NL msg deserialisation, but for the
only purpose of carrying attribute ids short constants to the
attribute parsers. Removing it will ease the simplification and
optimisation of the deserialisation code.

This patch also changes the types of constants declared in the
OpenVSwitch interface to reflect their usage in the ovs layer. All
command id constants are changed from int to byte, and all attribute
id constants are changed from int to short.

Change-Id: I494b211dad485f0dc211d79c9075abe2d9873c6a
Signed-off-by: hugobenichi <hugo@midokura.com>