Enhancements and unit tests for VtepDataClient
This contains unit test for most methods in the VtepDataClientImpl,
leaving out some cases that are pretty much boilerplate. It also
refactor the Guice configuration and introduces a couple of injectors
inside the VtepDataClientImpl to allow mocking from the unit test, and
also follow Guice's conventions a bit better.

Also refactored the VtepDataClientProvider moving the real
implementation to the brain module so that it can be used from there
without needing a dependency in the API.

Change-Id: I3edef51b56ff54296ce012e2d68dc5cbb28c3ecd
Fixes: MN-1814
Signed-off-by: Galo Navarro <galo@midokura.com>