Send a random UDP to migrate MAC between trunks
MAC migration invalidates all flows to the old port, to which
the migrated MAC has been mapped. Those flows are tagged by
br_fwd_mac.

Because flows FROM the old port are not invalidated, flooding
an ARP request from Midonet and expecting the ARP reply to
migrate the MAC may not work. The ARP reply flow is not to,
but from the port.

Fix MN-1819

Change-Id: I7b9c7e0f8c48004cb8a896920d369b708caf4306
Signed-off-by: Takayuki Usui <takayuki@midokura.com>