VxGW: Consolidate VTEP state
Introduces a new step before starting to watch a bridge that performs
the config. consolidation between Midonet and a VTEP by taking the
binding information from Midonet bridges and ensuring that the data in
the OVSDB is consistent. This does *not* seed the existent MAC-Port
entries, which will be implemented in a separate patch.

Ref. MN-1827
Change-Id: Id39823d8a3f4d8c4d766be4424ca1098f0b72853
Signed-off-by: Galo Navarro <galo@midokura.com>