Fix the broken SG CRUD test
SG CRUD test randomly fails because the List comparison is
performed on lists that may differ in ordering.  Let
SecurityGroupRule implement Comparable so that it can be sorted
in a list.  An alternative solution would be to make the
List comparison in SecuritryGroup ignore ordering, but opted
this time to go with the other solution.

Change-Id: I4a61450e3539f71bae2e21881e0191cc6883ad7a
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>