VXGW: stop monitoring bridge on vxlan port removal.
A MidoVxLanPeer monitors all bridges associated to a given VTEP,
and forward any changes to the bridges' MAC-Port tables to the
VTEP. This patch adds monitoring of the bridges' vxlan ports, so
that if one of such ports is removed (i.e. the bridge is decoupled
from the VTEP), the the MidoVxLanPeer stops monitoring it for
changes in the MAC tables.

Ref: MN-1828

Change-Id: I3ddbea3f0052721e950713e6c5e74cfa2e1ca878
Signed-off-by: Ernest Artiaga <ernest@midokura.com>