Change log level of "duplicate dp flow add"
With the new packet pipeline architecture introduced in v1.4, flow
add requests sent to the dp that fail are no longer handled directly.

Instead, a packet that should have hit a flow in the dp will now be
handled by the corresponding wildcard flow at the beginning of the
packet pipeline. This will trigger an new flow add request, that if
it succeeds will "heal" the dp flow to wildcard flow mapping in the
flow manager. As a side effect of hte new request, the same dp flow
will be pushed to the flow manager.

This patch changes a log statement lvli, since trying to add a
duplicated dp flow in the flow manager is now part of this self-
healing mechanism and is not an incorrect operation anymore.

Change-Id: I96f1bba09c979a995055f031b379bf811b13cefd
Ref: MN-1779
Signed-off-by: hugobenichi <hugo@midokura.com>