No list boxing in handling of netlink answers
When handling netlink messages answering pended requests, the
NetlinkRequest objects would maintain a list of ByteBuffers and
would wait until completion before applying the deserialisation
function given by the request client. This was causing an extra
List boxing on all single-answer requests, and was causing extra
ByteBuffer allocations and copying on all multi-answers requests.

This patch removes this boxing by specialising the NetlinkRequest
class for both cases. In the multi-answer case, the specialized
NetlinkRequest class is building a Set of user objects using the
given deserialisation function as ByteBuffers are filled with data
from the Netlink socket and passed to the NetlinkRequest. In the
single-answer case, the translation function is applied immediately
when the ByteBuffer containing the answer is passed to the
NetlinkRequest handler. For single-answer requests that only ask for
an ACK, null will be passed to the deserialisation function.

This also allows to share the same deserialisation function for
both enumerate and get requests and reduces deserialisation code
duplication.

Because the data in the reply ByteBuffer is now comsumed immediately
by the handlers, it becomes also possible to remove all slicing and
copying operations of the reply buffer.

Signed-off-by: hugobenichi <hugo@midokura.com>
Change-Id: I0530595d0864431e481772cbe73d3daafd911b24