Getter of vxlan endpoints for ports in dataclient
The VtepClusterClient contained some code to feed MacPort entries in
exterior ports to a VTEP. This will be needed shortly in the Brain for
consolidation of VTEP-Mido state. This commit moves the code that
figures out the set of (port, vxtunnel-endpoint) pairs into the
LocalDataClient so that it can be used from the Brain later.

Ref.: MN-1827
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I91c51b3e7f25b975fa6d428ca69b97bc3370d711