Ensures userspace packets don't create flows
This patch ensures that packets with the FlowActionUserspace reason
don't create wildcard flows, because we don't query the wildcard flow
table for these types of packets. It also ensures we don't try to
create additional DP flows, as one already exists in the datapath,
which originated this packet.

Note that we execute the packet, but the DP also executed the packet.
This means that packets will currently be duplicated.

Change-Id: I2a741c3a8b4d930045a208373cb99ea40678556c
Signed-off-by: Duarte Nunes <duarte@midokura.com>