VxGW: push MACs from VTEP to Mido on startup
Upon startup, the VxLanGatewayService will pull the existing entries in
the Ucast_Macs_Local table of the VTEP and push them to the Midonet
storage on the appropriate bridges and ports, following the same code
path as during dynamic sychronization (through the existing observable,
etc.).

This includes some changes to the connection logic preparing for the
configuration consolidation that will be implemented soon after. In
order to be sure that we don't miss any mac-port entries from the VTEP,
once both vxlan peers are connected and stable, we trigger an
"advertise" of all the macs in the VTEP.

Ref. MN-1827
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I544ff81630c3dfca7aa7b15c256cb628736711c3