Ensures wcmatch input port UUID is always unset
This patch fixes a bug where we weren't clearing the input port UUID
of the WildcardMatch of a simulation that resulted in a drop flow.
This meant that the WildcardMatch added to the flow table would have a
different hash then the WildardMatch of new packets, which don't have
the input port UUID set when we query the WildcardFlow table.

This bug affected mostly ICMP echos, which are always handled my
Midolman.

Ref MN-2327

Change-Id: Id9987fb5e15d34f82b1eef907ff48098e6ee66ea
Signed-off-by: Duarte Nunes <duarte@midokura.com>