Distinguishing token policies from DpPort types
The TokenBucketPolicy was linking DpPorts to the HTB making
assumptions about the usage of a DpPort based on its type. This
was causing problems to distinguish between the vxlan tunnel port
used for overlay traffic from the one used for vtep traffic.

This patch introduces a simple enum-like algebraic type which is
given as an argument to createAndHookPort() in the upcall connection
manager. This value is then used by the TokenBucketPolicy for linking.

Change-Id: Id2df3c526a11cf1522b02b0d5d6927e3c0ec212e
Signed-off-by: hugobenichi <hugo@midokura.com>