Resolve deprecation in RoutesTrie
RoutesTrie had a method to check if two ipv4 addresses given as 32bits
int had matching prefix. The method was tagged as deprecated and to be
replaced by a similar function in IPv4Subnet.

To resolve about 20 compilation warnings in midolman/src/test, this
patch moves the implementation and test of this method in the
IPv4Subnet class. The "raw" method using ints is still exposed to avoid
object creations in places where it matters.

Change-Id: Ib1a07d37f9b827e627b0345c35745056afb315ea
Signed-off-by: hugobenichi <hugo@midokura.com>