VXGW: Bug fix: missing FloodingProxyWeight from host lists.
The FloodingProxyWeight was not retrieved when obtaining the host
information via the REST API method to obtain a list of hosts
(opposed to retrieve a host individually). This patch adds the
proper APPLICATION_HOST_COLLECTION_JSON_V2 tag to the list
method, so that the missing property is correctly encoded and
retrieved.

Ref: MN-2394

Change-Id: I817f626bcd9ae5bd392cc3b9ce9c986743e05958
Signed-off-by: Ernest Artiaga <ernest@midokura.com>