Removing IntIPv4 from PathBuilder
IntIPv4 was used in PathBuilder to create ZK node name with an
alternative toString method for making valid URI path from cidr
notations. Replacing blindly IntIPv4 with IPv4Subnet would break
this invariant because the toString method of IPv4Subnet was
outputting "standard" cidr strings like 172.10.0.0/16. In this
patch, the toZkString() is used instead to solve this issue.

Change-Id: I3a6a2ccd942b8de0f20f6ffe5b923d98c92604de
Signed-off-by: hugobenichi <hugo@midokura.com>