Fix the validation for the weight of PoolMembers
This patch fixes the bug of the handler of the PoolMember sub resources.
It could take the 0 weight and pass it to ZooKeeper as is without the
validation but this patch prevents it and force users to give the weight
which is larger than 1. If the invalid nubmers, i.e., the negative
numbers or zero, the server returns 400 Bad Request as the response.

In addition to that, the weight defaults to 1 instead of 0 with this
patch.

This fixes MN-2401.

Change-Id: I3adde84bc1e02c6d4d573cd1b1ec3f583918ca8d
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>