Licenses initialized after starting the REST API
The patch-set creates a new license service that is initialized
after the REST API. The service reads the current list of
licenses from ZooKeeper upon start and saves the licenses
to ZooKeeper when stopped.

This ensures the licenses are initialized after the environment
setup that creates the directory structure in ZooKeeper.

Fixes: MN-2424

Change-Id: If0238a7986163aa155d0fdcd1583f53429c19657
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>