hashCode() & equals() cleanup in FlowKey classes
Also, for the FlowKey classes whose hashcodes were lazily computed,
this patch instead eagerly computes the hashCode at ctor invocation
or at deserialisation. The rational is that once a FlowKey is created,
the probability that its hashcode gets computed is alsmot 1. Therefore,
better do it when the data is hot in the cache than later.

Change-Id: I1aa1c1b43eba9b1c2f981cdcd85a677aa9d405b9
Signed-off-by: hugobenichi <hugo@midokura.com>