VxGW: support VTEP ownership in DataClient
Adds a method in the DataClient that allows trying to take ownership of
a VTEP. This will be used from the VxLanGatewayService to ensure that
only one node at a time can be syncing MACs from/to the VTEP.

The VTEP reservation is done using ephemeral ZK nodes. A node takes
ownership by writing an ephemeral child in the VTEP node with data
containing the id. If it doesn't exist, then we try to write the id from
the request. If an owner already exists, or we lose the race when
creating the child node and someone else took ownership, then we back
off and report the actual owner.

Ref.: MN-1830
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I3c616f9baea407287d858855c08925739c5ff243