VXGW: L2 flooding proxy
Adds a flooding proxy service to the VXLAN gateway, which is
responsible for configuring a flooding proxy host in a physical
VTEP for every logical switch. The service monitors VTEPS, bridges
and hosts to update the configuration of the existing flooding
proxies.

The service maintains an internal state for all entities involved,
in order to allow a fast determination of what entities are
affected by a given update (creation/change/deletion), and whether
it needs to reconfigure the flooding proxy for a particular logical
switch.

Ref: MN-1693

Change-Id: I396b34b8c4c60c4662cd75d7c00d2a1a56bf7328
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>