Update ZooKeeper entity monitors with generic identifiers
Allows the possibility of using data types other than UUID
for the identifiers of monitored entities in ZooKeepers. This
is required at least for the monitoring of VTEP entities, which
use IPv4Addr rather than UUID as identifier types.

Updates the EntityIdSetMonitor class to notify a single stream
of events related to the creation and deletion of entities.
Upon demand, this class can also synchronously notify the current
list of entities.

Creates a generic VXGW device monitor class that monitors
changes from a particular class of entities.

Change-Id: I4cc5e2036b88ed858c75bd3a037a3103877a6e3c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
Signed-off-by: Galo Navarro <galo@midokura.com>