No ByteBuffer slicing in Packet deserialisation
Replacing slicing by saving the buffer bounds on the stack and
restoring the bounds after the submethod has consumed the slice
and returned.

Implicitly, this means the buffer will keep its original endianess
as set at the beginning of a Packet deserialisation. It is set to
BigEndian in Ethernet.java, and restored to system order in the ovs
layer in Packet.java. This remains coherent with the behavior before.

Change-Id: Ia76ef1f6f12147e5a10f4057af30c8748fac8766
Signed-off-by: hugobenichi <hugo@midokura.com>