VXGW: Vtep creation/deletion detection by VxLanGatewayService
This patch is a preparation for the tasks described in MN-1593,
namely the detection and management of newly appeared VTEPs.
The patch reorganizes the VxLanGatewayService and the
VxLanGwBroker code to make easier to process individual VTEPs.
It introduces the use of a VTEP monitor to detect VTEP creations
and deletions and update the necessary data structures in the
VxLanGatewayService.

Ref: MN-1593

Change-Id: I0fcaf5ec9aa55c56e6d688a9d0a82056bfad9c72
Signed-off-by: Ernest Artiaga <ernest@midokura.com>