Correctly generates ICMP flow actions
This patch fixes a bug where two ICMP data arrays were incorrectly
being tested for equality. As a consequence, we would add an ICMP flow
action which we would then not translate as no ICMP field was marked
as seen in the WildcardMatch. We would fail to install a proper DP
flow, so matching packets would always come to userspace.

The patch also removes the special UserspaceFlowActionTranslator in
favour of doing this translation in the FlowTranslator. Relevant unit
tests were added.

Ref MN-2484
Signed-off-by: Duarte Nunes <duarte@midokura.com>

Change-Id: I2d4bf606147160b8c730d004218b73af16a86bb3