Correctly generates ICMP flow actions
This patch fixes a bug where two ICMP data arrays were incorrectly
being tested for equality. As a consequence, we would add an ICMP flow
action which we would then not translate as no ICMP field was marked
as seen in the WildcardMatch. We would fail to install a proper DP
flow, so matching packets would always come to userspace.

Ref MN-2484
Signed-off-by: Duarte Nunes <duarte@midokura.com>

Change-Id: I0b99d772a384b850b3d5a4b33339d4ef7d1d6716