Fixes a NPE when setting the input port to null
This patch fixes a NPE that happened when setting the input port
UUID  to null. This results in a drop flow being installed, which,
when debug logs are enabled, prints the WildardMatch. In there, the
input port UUID is in the used fields set, but it's null, so the
toString call on it when results in the NPE being thrown.

Change-Id: Ia590b193f5aa828c19d2dcbac4c60e2cfbc1b049
Signed-off-by: Duarte Nunes <duarte@midokura.com>