Support for exists watches in Zookeeper Mock Directory
This patch allows ZK clients to set an exists watch on a
yet non existing ZK node, and receive a notification when
the node has been created. For this purpose, it adds a
global list of exists watches in the ZK mock directory.
Watches are added to this list whenever the client calls
the exists() method with a non-null watche, and the ZK
path does not exist.

Change-Id: I728154260252c0addd561cc48b14a10e4438d735
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>