Remove old Monitoring feature.
The old cassandra-based monitoring feature is not being used at the
moment because it's unable to scale in a production environment (mostly
due to the Cassandra schema). Fixing this alone will require an almost
complete rewrite. Additionally, this code adds dependencies to Cassandra
and the Metrics that increase the work required to bring up client
versions on these two libraries.

Change-Id: I38de316741bf0f663d4a5245074e56728e38a8bf
Fixes: MN-2554
Signed-off-by: Galo Navarro <galo@midokura.com>