VxGW: fix race condition on first binding
When a bridge is first bound to a VTEP, we have a race between the API
and the VxGWService consolidation component for the creation of the
relevant rows in the OVSDB.

This patch solves this by ensuring that the first binding setup in the
OVSDB is completely handled by the VxGWService itself. The API will just
write the binding to ZK, but not to the VTEP. On subsequent bindings on
the same network, the entry will be written both to ZK and the VTEP.
However, the write to the VTEP will never be required to be successful
for the overall operation to succeed. The API relies thus on the
VxGwService to consolidate state.

This patch involves a significant refactor on the addbindings code,
mostly dedicated to splitting the case where the binding is the first
one on a given network, and when it's not. This is because the latter
does not require VTEP reachability, the former does.

Also, makes sure to clean the VxlanPort on the bridge for the first
binding if a conflict is detected. Covered this with a test case.

Change-Id: I05dd3afba0dd627045035112a6339e34f9778fc8
Fixes: MN-2570
Signed-off-by: Galo Navarro <galo@midokura.com>