MN-2573 fix
This patch allows using different IPs for vtep management and tunnel end points. It introduces a new type of tunnel-zone (vtep) that is to be used
for vtep tunnels, while being ignored by the inter-agent communication.

This may have impact on other parts of midonet, so further testing
would be necessary.

Ref: MN-2573

Change-Id: I495eb40fa3956b75ba7b1b5f56437d854e52d65d
Signed-off-by: Galo Navarro <galo@midokura.com>