Pauseless L4: introduce the connectionHash
L4 pauseless simulation design, and proper deduplication in general,
require that packets belonging to the same connection, and not micro-flow,
be processed serially and on the same DDA thread.

This patch makes that possible by adding a new connectionHash() method to
FlowMatch and FlowKey. Only keys that have fields that belong to the connection
5-tuple l3/l4 return a value for this hash. And they calculate it using
exclusively the fields in the 5-tuple.

Change-Id: I23e2012cb09cdf7204fd1b1d55d518078b47bc78
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>