Do not take ownership of vteps
We generate a random uuid and store an ephemeral node in ZK when
starting the VXGW service. If the API is restarted, we generate a new
one, but since ZK hasn't expired the node yet and we're unable to detect
it, we end up not taking ownership of the (free) VTEP. The workaround is
to wait for 30s before restarting the API.

This patch removes the ownership of the VTEP, relying on the current
requirement that only 1 VxGW service can be active at any point in time.

Ref. MN-2572
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I2bd5f56433b8b9dbdba30d073fda06fa5897426f