VxGW: clean up unused LogicalSwitches on VTEP
When a binding is removed, the VTEP might not be reachable so the API
isn't unable to clean the binding data. This patch enhances the VXGW
consolidation phase so upon connection to a VTEP, all the old Logical
Switches that do not have a matching binding stored in Midonet will be
removed from the VTEP.

With tests.

Change-Id: Ib0e98d0c644786327a4aca832663227c93c42aed
Fixes: MN-2598
Signed-off-by: Galo Navarro <galo@midokura.com>