VXGW: Support for VTEP failover
This patch adds a class for VTEP monitoring, which keeps track of
all existing VTEPs and their ownership. The monitor exposes a set of
observables, informing the VXGW service of the VTEP ownership changes:
acquiring the ownerhip of a VTEP (either newly created or which lost
the owner), and release (either when deleted, or when the owner has
been changed by a third party).

The VXGW uses the monitor observables to setup and release the VTEP
pairing within the VXLAN gateway broker. The VXGW service no longer
needs to enumerate the current list of VTEPs and attempt acquiring
ownership, as this will happen automatically through the VTEP
monitor.

Ref: MN-1593
Change-Id: I41f288ed8231e07be008add73527bb8d11d9cba1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>